feat: lazy start PS when it's binded but not started
This prevents problems with unexpected autostart of engine service on
some devices. PlaybackService may be started by system any time via bindSevice()
because it implements MediaBrowserService. But we should not start
engine service in such case. Solution is to start engine service only
when startService() was called (and onStartCommand() was executed as
result).